+++ gas/config/tc-mn10300.c
@@ -2325,7 +2325,7 @@ md_apply_fix (fixS * fixP, valueT * valP, segT seg)
 {
   char * fixpos = fixP->fx_where + fixP->fx_frag->fr_literal;
   int size = 0;
-  int value = (int) * valP;
+  int value = *valP;
 
   gas_assert (fixP->fx_r_type < BFD_RELOC_UNUSED);
